Show a helpful error message when a package.json file can't be parsed
Before this change, if there was an invalid package.json file, flow
would just error out with this unhelpful output:

`Fatal error: exception Parse_error.Error(_)`

Now, the user will be able to see the actual path to the file that
caused the error.

This is especially helpful because the package.json file might be hidden
deep in the node_modules directory structure. This actually happens in
real life with packages like `enhanced-resolve` that has an
`invalidPackageJson` test fixture.

The "real" fix here is to only parse package.json files that correspond
to actual imported code. This is just a small usability fix.

As a workaround, users should be able to add the appropriate entries to
the [ignore] section of the .flowconfig file. This fix will at least
help users discover which paths need to be ignored.

See #353