Allow `undefined` to be passed to optional params
There is already a flow from `VoidT` ~> `OptionalT`, but in the case of
function parameters, `multiflow` actually unwraps the `OptionalT`, so by
the argument flows into the parameter, flow isn't able to detect this
case.

I fixed this by adding a redundant case to `multiflow` instead of a
larger overhaul that would use the existing flow.

Note that functions with default argument values use the same mechanism
as optional parameters. The included test ensures that the default
expression type still flows correctly.

Fixes #437