import typeof
Summary: This adds the `import typeof` declaration.

This is distinct from `import type` in the same way that `var a: Foo` is distinct from `var a: typeof Foo`.
In other words if you think of `import type` as a way of making a local alias to a remote type, `import typeof` is way of making a local alias to a `typeof` of a remote type.

One example of when this is useful is if you use DI and wish to annotate that a function takes the class-type of some remotely defined class.

Reviewed By: @gabelevi

Differential Revision: D2175972