[Facepalm] Fix "flow is busy" during server initialization issue
Summary: `CommandUtils.connect_with_autostart` handles connecting the client
with the server, including the situation when the server doesn't respond. For
some reason, we had wrapped that function call with another layer of error
handling. So when initialization was taking awhile, we would freak out that we
couldn't talk to the server. So before if you started a server with `flow
status` then you would see something like

  Flow server launched for /path/to/root
  Spawned flow server (child pid=123456)
  Logs will go to /tmp/flow/zSpathzStozSroot.log
  The flow server will be ready in a moment.
  flow server still initializing. If it was just started this can take some time. Retrying... |
  Error: flow server is busyzing. If it was just started this can take some time. Retrying... -
  Error: flow server is busyzing. If it was just started this can take some time. Retrying... -
  Error: flow server is busyzing. If it was just started this can take some time. Retrying... |
  Out of retries, exiting!lizing. If it was just started this can take some time. Retrying... -

Now it just says that the server is initializing until it's done initializing

Reviewed By: @mroch

Differential Revision: D2178206