Fix 4.01 builds
Summary: It seems 4.01 doesn't like `include` in one sense or another (I've broken the build with similar issues in the past where an .mli file needs to be far more explicit about opens/includes in 4.01 than is necessary in 4.02).

https://travis-ci.org/facebook/flow/builds/67932557

Unfortunately I wasn't able to find a way to make 4.01 happy here such that `utils_js.ml` could properly extend `utils.ml` so that modules that use it could get both interfaces in one module (I scoured google and stack overflow, but couldn't find anything that worked short of creating a .mli file for utils_js.ml and repeating everything from utils.ml in it. Other suggestions are welcome here if anyone has any ideas).

It would be nice if we could just kill 4.01 support altogether since this seems to be turning into a trend, but I'm not sure if there's a good reason we still support it -- so in the meantime this will get the Travis build going again.

Reviewed By: @bhosmer

Differential Revision: D2182724