[easy] non-strict lookup of props in conditions
Summary: There's this common JS idiom of testing the existence of a property before use.
While Flow tries to refine the type of a property when tested, it would still
need that the property to exist (and have a type). Optional properties don't
always provide a clean solution, since sometimes the programmer chooses between
different shapes based on the existence of a property.

This diff blesses that idiom by making property lookups non-strict when
processing conditions.

Reviewed By: @mroch

Differential Revision: D2183403