use make_next_files to find libs and package.json
Summary: consolidating codepaths to a single way of finding files (`make_next_files`) so that we can update a single place to follow symlinks. following symlinks requires listing all of the files and checking to see if they're symlinks; if we find by name, `find` won't follow the symlinks (unless we use `-R` which we don't want to do for perf reasons).

this is miraculously exactly the same speed as the previous implementation.

as a minor side effect, we now allow lib files to have any of the valid flow extensions (`.js` and `.jsx` for now, but see https://github.com/facebook/flow/issues/499) instead of just `.js`, and ignore dot-files in lib directories like we do elsewhere.

Reviewed By: @avikchaudhuri

Differential Revision: D2188648