locs_of_trace
Summary: Another way to format trace info: instead
of a tree, this gives a sorted list of unique
locations, each with a concatenation of all
the descriptions of the reasons that appeared
in that position. Motivated by the idea that
currenty, most people just scan trace output
for positions rather than actually trying to
follow the story.

Note: this diff just adds the function, it
doesn't hook it up to anything. Gettng it
in so it gets carried along in Marshall's
upcoming move of trace stuff.

Reviewed By: @avikchaudhuri

Differential Revision: D2176026