Add support for scoped type variables
Before this commit, type params were passed around in a map in `convert`
and related functions, but we quickly "forgotten."

This commit adds a ton of plumbing to pass around this map, accumulating
entries as scopes pile on.

I also changed the name of this variable from `map` to
`type_params_map`, for readability and because the scope of this
variable is super broad now, so it should have a specific name.