sentinel properties
Summary: Support refinements based on sentinel property checks (obj.key === string).

There is a common JS idiom to simulate ML-like ADTs (disjoint unions)
but taking unions of object types, with one distinguished property (sentinel)
containing the "case name" and the rest containing differently-shaped payloads
based on the value of the sentinel property. Case analysis is then simulated
by checking the value of the sentinel property. To avoid spurious errors,
Flow needs to narrow down the union of object types to the object type with
the correct shape depending on what branch the code is in. This diff does that.

Some stuff that this diff needs to build up includes singleton types.
It also cleans up / renames "enums," which are related and have been used to
model singleton types indirectly (and somewhat awkwardly). Since singletons are
part of the language and enums are still experimental, it makes sense to model
singletons directly and not via enums.

Another thing this diff does is clean up how we handle binary tests. The only
case previously was instanceof, but sentinel properties add one other case.
Now, instead of having pairs of predicates for each binary test, we now have
a distinguished pair of predicates to handle binary operands and then one case
for each binary test.

Reviewed By: @bhosmer

Differential Revision: D2188305