Promise covariant param
Summary: Promise's type param can be covariant, given the
class's signature. We have no way of specifying
type param variance currently, but given the
pervasive and hidden use of the Promise type
in async/await, we decided it was worth hardcoding
this special case - the work for the general
feature is nontrivial, since it includes

* deciding on def-site vs. use-site variance
* validating variance specification against signature
* parsing annotation syntax

This diff hardcodes (def-site) covariance for
Promise's single type param. However it's not a
complete hack; it builds out the back end of a
def-site variance pipeline that will be useful
as part of the general solution, assuming we go
with def-site specification (not a sure thing).

Note: the hardcoding will apply, without validation,
to any `declare class` or `interface` named Promise
with a single type param, ours or a user's. Played
with putting it behind a type alias, but ran into
unexpected side effects from the indirection.

Added a test dir `type_param_variance`, which
should become a non-parody name once we have
the general functionality.

Reviewed By: @avikchaudhuri

Differential Revision: D2185549