[easy] complete transition of keys to union of literals
Summary: The notion of SingletonT was introduced in a previous diff.
In many cases KeysT could be modeled as union of SingletonT.
But some TODOs remained, e.g. React.PropTypes.oneOf was not moved over yet.

This diff completes the transition.
It also makes explicit one of the rules for ordering patterns in __flow.

Reviewed By: @bhosmer

Differential Revision: D2195198