enable esprima tests for iterators
Summary: flow supports for-of but the esprima tests were disabled.

one of the cases (`for (var x = 42 of list)`) fails in flow but passes in esprima-fb 15001.1.0-dev-harmony-fb. afaict, it's fixed in upstream esprima so we can delete the expected difference once we merge/upgrade.

also, turns out that the esprima test runner doesn't handle the case where we error but esprima doesn't. fixed that by making it generate diffs for the errors so that we can expect them.

Reviewed By: @gabelevi

Differential Revision: D2193570