recursively follow symlinks (again)
Summary: trying again to follow symlinks. previously, we tried to use `find -R`, but if it hits a cycle, `find`'s behavior is somewhat undefined. from the GNU find man page:

```
The POSIX standard requires that

	      The  find utility shall detect infinite loops; that is, entering
	      a previously visited directory that is an ancestor of  the  last
	      file  encountered.  When it detects an infinite loop, find shall
	      write a diagnostic message to standard error  and	 shall	either
	      recover its position in the hierarchy or terminate.
```

GNU find writes to stderr and exits code 1... it seems like it recovers and continues, but I don't really feel comfortable relying on it especially because of the error code. BSD find (OS X) also warns but exits code 0.

`find -R` also makes the symlinks more or less invisible... if you've got a symlink from `foo/` to `bar/`, `find -R` will return both `foo/file.js` and `bar/file.js`. we then have to `realpath` them both (or `stat` and compare inodes) to realize that they point at the same file. I think having to enumerate all of the files multiple times in `find`, plus deduping them in ocaml, ends up being a lot of extra work.

Instead, this diff modifies the file finder to iteratively follow symlinks manually. When `find` returns a symlink, we manually resolve it and if it is a directory, add it to a queue. When each `find` pass completes, we call `find` again with all of the newly found symlink targets, and so on until the queue stays empty.

Reviewed By: @bhosmer

Differential Revision: D2183502