[hack] Combine represent all kinds of abstract types as Tabstract in locl phase
Summary: Currently we have 3 notions of abstract types in the type checker.
  * Generics
  * Opaque Types
  * Expression Dependent Types

These types are handled very similiarily when type checking, but since they are different variants it is very easy to forget a branch. To fix this I combine all three kinds of abstract types into a variant ##Tabstract##, that consists of a constraint type it wraps and a new variant type ##abstract_kind##. The ##abstract_kind## can be examined when we want to write special case logic on a specific king of ##Tabstract##.

This cleans up the representation of expression dependent types which were either a Tgeneric or a Tabstract, and required examining a string to distinguish it from typical Tgeneric and Tabstract type. Adding another type that was very similar seemed error prone since I could forget places where I would need to pull the constraint type it wraps.

Reviewed By: @jwatzman

Differential Revision: D2177310