fix [ignore] of package.json
Summary: Module_js wasn't properly filtering on [ignore]
when it looked in the file system for package.json
files, so an attempt to read package info would
provoke an internal error rather than either
"module not found" or a fallback to a lib.
This diff adds the necessary check.

Also unbreaks the build from moving Loc out
to its own module.

Reviewed By: @mroch

Differential Revision: D2203750