[hack] Remove implements check for abstract ancestors
Summary: Before this diff, we would check that a given class was consistent with its
direct parents as well as all its abstract ancestors (even if they were grandparents).

The special-casing of abstract ancestors was originally
introduced in {D558649} in order to allow abstract classes to implement
interfaces without re-declaring all the members in an interface.
However, this seems redundant now; we implicitly mark all methods on an
interface as abstract, so we know that an abstract class does not need
to implement them.

Avoiding these checks cuts the time we spend in the Type-check phase by ~1/3.
We end up only checking the direct parents of a class for consistency.

This change would be completely sound if the class implements checks were
transitive. However, this is not the case because type hints are optional in partial
mode. Thus, if a given class was inconsistent with its grandparent class, but the parent
class lacks type annotations, we would not raise an error. However, note that we
already miss out this case if the parent class is not abstract, so while this commit
makes us slightly more unsound, it also makes us more consistent in our unsoundness.

To wit, some examples:

This is currently not an error because `Bar` is concrete:

```
class Foo {
  public function f(int $x) {}
}
class Bar extends Foo {
  public function f($x) {}
}
class Baz extends Bar {
  public function f(string $x) {} // inconsistent with Foo::f
}
```

This is currently an error because `Bar` is abstract, but will not be an
error after this diff is applied:

```
abstract class Foo {
  public function f(int $x) {}
}
abstract class Bar extends Foo {
  public function f($x) {}
}
class Baz extends Bar {
  public function f(string $x) {} // inconsistent with Foo::f
}
```

Reviewed By: @elgenie

Differential Revision: D2052045