no implicit typeapp
Summary: Previously, Flow would see annotations : A (where A is a polymorphic class)
and implicitly take them to mean : A<X> (where X is an infered type variable).
This caused confusing errors, especially in React code where return types of
render methods would be declared ReactElement, while their bodies would return
different ReactElements, causing the props of the different React classes to be
unified. This points to a design mistake.

This diff reports an error when type args are found missing in annotations.
Note that implicit instantiation still (needs to) take place when using
polymporphic classes as runtime values, e.g. for new, extends, static calls, etc.

(Another approach, not taken in this diff, would be to take : A to implicitly
mean : A<any>. While this choice often agrees with intuition, it would imply two
different strategies for implicit instantiation, depending on use:
any for types, X for runtime values.)

Reviewed By: @bhosmer

Differential Revision: D2189397