[hack] Shapes::idx helper function for accessing optional fields
Summary: Optional shape fields require ways to access them that take into account that it might not set - "[..]" subscript will just emit a warning if it's not, without a way to test or avoid the warning. It seems that minimal "standard library" for shapes would consist of:

  shape_idx
  shape_key_exists
  shape_key_unset
  shape_array_cast

The naming seems to be slight point of contention. Other alternatives include static utility class:

  Shapes::idx
  Shapes::key_exists
  Shapes::key_unset
  Shapes::array_cast

or overloading existing array functions:

  idx
  array_key_exists
  unset
  (array)

The first choice seems consistent with current standard library (even if we don't like it) and makes shapes first-class citizens of language, which they are. The last one is blurring the line between shapes and arrays too much for my liking. Static utility class is a way that we might still pursue in addition to having top level functions, if and when we decide to push the rest of stdlib that way. For now  I'm going to vote for the first version.

UPDATE: Nope, changed my mind, Shapes::idx it is.

This diff adds first of those functions, Shapes::idx. Most of the complexity is due to unresolved types - to make it useful we want to error when accessing field that is never set, but in case of unresolved we need to check all variants if any of them has the key.

Reviewed By: @dlreeves, @jwatzman

Differential Revision: D2116812