add --verbose-indent option
Summary: A variant of `--verbose` that indents its output based on the depth of the trace.

I find this really helpful when looking at verbose output -- it helps see the recursion, so you can follow which rule caused something to decompose into the next rule, vs when it's walking back up the stack.

`--verbose-indent` can be used instead of or in addition to `--verbose`. Didn't call it `--indent` to keep that option available for some theoretical non-debugging use case.

Reviewed By: @bhosmer

Differential Revision: D2218749