[hack] Make the expression $this a dependent type
Summary: The variable ##$this## is a special variable that cannot be assigned to at
runtime and always refers to the late static bound type. By making it an
expression dependent type it allows aliases of ##$this## to also be compatible
with the late bound type.

I also cleaned up the error message a bit, so we don't bother explaining
expression dependent types if the error is unrelated to the dependent type.

Depends on D2203422

Reviewed By: @jwatzman

Differential Revision: D2203497