[hack] Make 'new class_id()' produce expression dependent type
Summary: When we create a new object using ##new class_id()## we know the exact late bound
type of the result. For instance ##new static()## create the late static bound
type of enclosing class (which will be compatible withh ##$this##). ##new A()##
creates a new instance of ##A## that we know is exactly that type (not a
subtype of ##A##).

This reveals a few bugs in the type checker where we are not handling abstract
types properly. I attempted to fix as many as showed up in our code base. There
is one more involved bug around creating nested ##Tunresolved## types. This
ironically causes us to lose the ##Tunresolved## in some cases causing errors
when we unify an invariant type parameter with an abstract type. To produce
this error is rather involved and there are only a couple examples of this
present in our code base so I think its safe to land this.

Depends on D2203497

Reviewed By: @dabek

Differential Revision: D2204210