[hack] Pass nenv through to `--convert` mode
Summary: Due to the unnamed body refactor, we may run naming during
typing, meaning we need the naming env around. In particular, since we
already do a typecheck run before we go into the `--convert` bits, we
should be able to assume no type errors, so we never set up error
accumulation, and will crash if we hit an error. This is a good thing,
except we need to thread through the nenv to make sure we don't hit
naming errors on undefined classes.

Both "need to set up error accumulation" and "unnamed bodies" are new
since we last sigificantly used convert mode, which is why we just now
noticed, when some external folks are trying to more extensively use
this mode.

Reviewed By: @dlreeves

Differential Revision: D2251195