Allow '*-/' to mark the end of a block comment inside a flow-comment
Summary: Currently Babel doesn't have a good option for commentizing something like:

```javascript
type F = /* test */ number
```

(at the moment it naively/invalidly transforms it to `/*::type F = // test number */` -- not accounting for single-line block comments)

So an idea that I suggested (and @​hzoo implemented) is one that allows Babel to convert a legit block comment's end-token to `*-/` -- which fixes the resulting raw JS source. However, we still want Flow to be able to understand this when parsing the flow-comment, so I've special-cased it in the Flow lexer here when we're inside a `/*::` flow-comment.

See here: https://github.com/babel-plugins/babel-plugin-flow-comments/pull/6#issuecomment-122709012 (and then https://github.com/babel-plugins/babel-plugin-flow-comments/pull/13 for the actual PR)
Closes https://github.com/facebook/flow/pull/647

Reviewed By: @gabelevi

Differential Revision: D2259564

Pulled By: @jeffmo