[hack] Fully qualify gena/genva/gen_array_rec special cases
Summary: Unlike the functions that we're treating like magical language
constructs, such as fun and inst_meth, these are real functions that we
just special case. We need to make sure they get fully qualified before
special casing them -- in particular, this makes it so that @fredemmott
can easily special case `\HH\Asio\va` etc.

The wisdom of leaving anything out of this is questionable, we may want
to give `fun`, `inst_meth`, etc the same treatment, but that's a
somewhat more involved change.

Reviewed By: @fredemmott

Differential Revision: D2262014