[hack] Error on overriding concrete function with an abstract one
Summary: This is a function counterpart of {D2237726}. I wanted to re-use the same error, but ran into the issue of not having function name available at the error callsite.

I considered threading it through, but removing it seems more consistent with other errors we have - we don't mention function names anywhere, we just point to them using position.

It also revealed a bug in our constructor inheritance - we have the logic that when having to arbitrarily choose between abstract and concrete method we will choose the concrete one, but we weren't doing the same for constrcutors, which apparently didn't matter, but it matterns now - without it, the new inherited_constructor.php test case would fail.

Reviewed By: @jwatzman

Differential Revision: D2253618