Declare overloaded versions of fs.readFile and fs.readFileSync
If no options are provided, data is a Buffer.
If options are provided without an encoding set, data is a Buffer.
If options are provided and an encoding is specified, data is a string.

If the second parameter is a string, data is a Buffer. This is the "old"
API, so it's undocumented, but it is still supported and does the right
thing.