[hack] Don't `sleep` waiting for a connection
Summary: Our retry logic before this diff is pretty dumb: if a connect
fails, do a `sleep 1` and try again. This is pretty painful for small
projects, where the startup time is often just slightly north of 1
second -- we end up waiting for 2 full seconds when 1.1 would have done
it.

We can actually be a bunch smarter, since the kernel will deal with
blocking and unblocking us for us, since reading/writing to a socket
with no one on the other end will block until they accept and read.

To do this, we need to make sure that the server actually starts a
server on the socket, even though it isn't listening -- our existing
`SIGALARM` timeouts in the client will deal with the rest.

One slight wrinkle with the way this happens is that, ideally, the write
on the client would be what would block. Unfortunately, it's not -- the
write gets buffered even though the server has not accepted the
connection yet, and the client blocks on the read. Not a huge deal, just
the server gets a few reads and then a broken pipe when it tries to
respond with its version. This makes D2271043 even more important, so
that it doesn't spew in the logs a bunch with expected broken pipe
errors after every startup.

Reviewed By: @int3

Differential Revision: D2271092