[hack] Make the experimental emitter a standalone binary
Summary: Make the experimental emitter be a standalone binary. This
gives us a way to run the emitter that is disconnected from
hh_single_type_check's roll as a test harness. hh_emitter does not run
typechecking and ignores naming errors so that it can be used to emit
files that would not pass checks when considered in isolation (that
is, ~all files in a real web project).

Reviewed By: @int3

Differential Revision: D2275707