[hack] Prohibit assigning to locals when there is another unsequenced access.
Summary: Prohibit assigning to locals when there is another unsequenced access.

This disallows "conflicting" accesses to local variables, where two accesses conflict iff at least one of them is a write and they are not "sequenced". We consider most constructs to sequence their subexpressions with the follow exceptions: non short-circuiting binops, indexing, and array key/value pairs. Assignment sequences the assignments to the LHS variables with respect to *reads* on the LHS but not with respect to writes (so '$x = $x + 1' is permitted but not '$x = $x++').

Since a common idiom in www seems to be to assign to unused variables as a form of documentation, we suppress the check for variables that are never read from.

This solves an unsoundness bug arising from code like '$x + f($x = $z);' in which the assignment to $x can change its type but the typechecker won't notice and I think makes strict Hack programs unable to distinguish their evaluation order from left-to-right. (References can throw a wrench in the works there.)

Reviewed By: @int3, @elgenie

Differential Revision: D2214757