[hack] Lint: Raise warnings on usage of namespaces
Summary: This required quite a bit of refactoring to allow
linting_service to handle parsing as well. We can't detect and raise
these warnings on the named AST because we've already flattened those
namespaces during parsing. (While we could check for class / function
names with more than one slash in them, we would not be able to obtain
the position of the `namespace` declaration.)

Reviewed By: @dlreeves

Differential Revision: D2279819